requirements.txt
chromadb==1.0.15
groq==0.30.0
langchain==0.3.27
langchain_community==0.3.27
langchain_ollama==0.3.6
minio==7.2.15
numpy==2.3.2
pandas==2.3.1
pymongo==4.13.2
python-dotenv==1.1.1
PyYAML==6.0.2
PyYAML==6.0.2
streamlit==1.47.0
streamlit_pdf_viewer==0.0.26
Werkzeug==3.1.3
google-generativeai==0.8.5
pypdf==5.9.0
